Move MemoryWorldDatabase to ../
